fix: Expose `pdb.agg(.., solve_mvcc=bool)` argument in all cases (#4416)
The `pdb.agg` `solve_mvcc` argument was only exposed on window
functions, but is also helpful on aggregate functions in general.

Additionally, the new argument had not been included in upgrade scripts,
and so was not exposed to consumers who had upgraded.

To allow for higher performance aggregates in cases where users don't
need mvcc accuracy.
