+++ docs/documentation/aggregates/overview.mdx
@@ -61,6 +61,44 @@ WHERE category === 'electronics';
 (1 row)
 ```
 
+## Performance Optimization
+
+On every query, ParadeDB runs checks to ensure that deleted or updated-away rows are not factored into the result set.
+
+If your table is not frequently updated or you can tolerate an approximate result, the performance of aggregate queries can be improved by disabling these visibility checks.
+To do so, set the second argument of `pdb.agg` to `false`.
+
+<CodeGroup>
+```sql SQL
+SELECT pdb.agg('{"value_count": {"field": "id"}}', false)
+FROM mock_items
+WHERE description ||| 'running shoes';
+```
+
+```python Django
+from paradedb import Agg, Match, ParadeDB
+
+MockItem.objects.filter(
+    description=ParadeDB(Match('running shoes', operator='OR'))
+).aggregate(
+    agg=Agg('{"value_count": {"field": "id"}}', exact=False)
+)
+```
+
+```ruby Rails
+MockItem.search(:description)
+        .matching_any("running shoes")
+        .facets_agg(exact: false, agg: ParadeDB::Aggregations.value_count(:id))
+```
+
+</CodeGroup>
+
+Disabling this check can improve query times by 2-4x in some cases (at the expense of correctness).
+
+<Note>
+If a single query contains multiple `pdb.agg` calls, all of them must use the same visibility setting (either all `true` or all `false`).
+</Note>
+
 ## JSON Fields
 
 If `metadata` is a JSON field with key `color`, use `metadata.color` as the field name:
