+++ pg_search/tests/pg_regress/expected/custom-agg.out
@@ -2192,13 +2192,18 @@ ORDER BY id DESC LIMIT 3;
   1 | Test item one   | A        |   100 | {"value": 200.0} | {"buckets": [{"key": "A", "doc_count": 2}, {"key": "B", "doc_count": 1}], "sum_other_doc_count": 0, "doc_count_error_upper_bound": 0}
 (3 rows)
 
--- Test 68: pdb.agg() in GROUP BY context with solve_mvcc = false should ERROR
--- solve_mvcc=false is only allowed in TopN (window function) context
+-- Test 68: pdb.agg() in GROUP BY context with solve_mvcc = false should work
 SELECT category, pdb.agg('{"avg": {"field": "value"}}'::jsonb, false)
 FROM mvcc_test
 WHERE description @@@ 'test'
 GROUP BY category;
-ERROR:  pdb.agg() with solve_mvcc=false is only supported in window function context (with OVER clause). GROUP BY aggregates always use MVCC filtering for correctness. Remove the second argument or use solve_mvcc=true.
+ category |       agg        
+----------+------------------
+ A        | {"value": 150.0}
+ B        | {"value": 350.0}
+ C        | {"value": 500.0}
+(3 rows)
+
 -- Test 68b: pdb.agg() in GROUP BY context with solve_mvcc = true should work
 EXPLAIN (FORMAT TEXT, COSTS OFF, TIMING OFF, VERBOSE)
 SELECT category, pdb.agg('{"avg": {"field": "value"}}'::jsonb, true)
